channels/discord/requirements.txt
discord.py>=2.3.0
httpx>=0.27.0
python-dotenv>=1.0.0
